+++ brane-job/src/worker.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    31 Oct 2022, 11:21:14
 //  Last edited:
-//    07 Mar 2024, 11:59:55
+//    01 May 2024, 10:39:39
 //  Auto updated?
 //    Yes
 //
@@ -67,8 +67,8 @@ use specifications::profiling::{ProfileReport, ProfileScopeHandle};
 use specifications::registering::DownloadAssetRequest;
 use specifications::version::Version;
 use specifications::working::{
-    CheckReply, CheckTaskRequest, CheckWorkflowRequest, CommitReply, CommitRequest, ExecuteReply, ExecuteRequest, JobService, PreprocessKind,
-    PreprocessReply, PreprocessRequest, TaskStatus, TransferRegistryTar,
+    CheckReply, CheckTaskRequest, CheckWorkflowRequest, CommitReply, CommitRequest, ExecuteReply, ExecuteRequest, JobService, PreprocessReply,
+    PreprocessRequest, TaskStatus,
 };
 use tokio::fs as tfs;
 use tokio::io::AsyncWriteExt;