+++ brane-shr/src/lib.rs
@@ -1,19 +1,19 @@
 //  LIB.rs
 //    by Lut99
-// 
+//
 //  Created:
 //    30 Sep 2022, 16:21:24
 //  Last edited:
-//    06 Jun 2023, 18:38:34
+//    01 May 2024, 10:44:25
 //  Auto updated?
 //    Yes
-// 
+//
 //  Description:
 //!   The `brane-shr` crate defines common functions and other tools used
 //!   throughout the framework. This crate differs from the
 //!   `specifications` crate in that it does not aim to standerdize
 //!   things, but rather just provide a common implementation.
-// 
+//
 
 // Declare some modules
 pub mod errors;
@@ -21,5 +21,5 @@ pub mod formatters;
 pub mod fs;
 pub mod input;
 pub mod jobs;
-pub mod kafka;
+// pub mod kafka;
 pub mod utilities;