+++ specifications/src/lib.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    07 Jun 2023, 16:22:04
 //  Last edited:
-//    07 Feb 2024, 11:55:02
+//    01 May 2024, 10:11:07
 //  Auto updated?
 //    Yes
 //
@@ -24,6 +24,7 @@ pub mod container;
 pub mod data;
 pub mod driving;
 pub mod errors;
+pub mod os;
 pub mod package;
 pub mod planning;
 pub mod policy;